encapsulate filters
